Change dealing logic to auto-deal with any number of cards.
This also changes the dealing to take place in pairs, at least as long as there are enough cards to deal an entire round in pairs. After that, singles until they run out. This way if the cards aren't nicely 'even', at least the Beggar always gets a card less (and optionally the guy(s) before him)